Added support for variables in the query
Variables in the query are marked by '?' or '$'. The marker is not part
of the variable name, so `$var` and `?var` represent the exact same
variable.

Also fixed a style issue with the error message. According to the
documentation for the error function the convention is that error
messages start with a capital letter but *do not* end with a period.