Fix indenting wrong line when using ;
sparql-mode.el (sparql-indent-line): `save-excursion' closing
parenthesis was in the wrong place.

What happens inside the save-excursion is that we go backward 1 line to
check if there is a ; on that line. If it finds one it sets the
indent-column to the beginning of the second symbol right before the ;,
but instead of resetting the point back to the previous position we
would keep the position in the previous line and indent that line
instead.