Fix comment and URL syntax highlighting when re-fontifying
sparql-mode.el (sparql-keywords): Stop overriding other faces(, like
the comment face). Add C-j as end-of-comment.

sparql-mode.el (sparql-syntax-propertize-function): Make
syntax-propertize-function skip URLs, but if we see a # we add it as a
beginning of comment.

These changes fixes a bug where the fontification of comments didn't
work properly. If you entered a URL with a #, but didn't close the <>,
the line would be given the comment face until you moved to the end of
the line and force a face update by adding a space or similar. We used
to rely on the fontification to detect if we should add the syntax
property, but it is easier to just ignore the parts that shouldn't be
fontified with comment all together instead.