ob-sparql.el add interop with org-link #+link: syntax for prefixes
If #+link: keywords are defined in a file or if org-link-abbrev-alist
has settings, use the values defined there to compact the full uris
that are returned from a query. Useful for making it possible to see
whole rows of results when there are long uris.

org-babel-sparql--current-curies is needed because table formatting is
done in a separate buffer so we use the dynamic variable to pass the
value between buffers